dynamixel_webapp/requirements.txt
Flask>=3.0.0
Flask-SocketIO>=5.3.0
dynamixel-sdk>=3.7.0
pyserial>=3.5
